+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - **BREAKING CHANGE**: broken manifests will cause a hard error again
 - Added `tool.tytanic.tests` manifest config option for configuring the unit test root
 - Added `tool.tytanic.default` manifest config section for configuring defaults, contains options for `ppi`, `dir`, `max-deviations` and `max-delta`.
+- Added `ppi`, `dir`, `max-deviations` and `max-delta` annotations.
 
 ## Fixes
 - Don't remove references of persistent tests on `tt util clean`